Rename tensor types and move to own module, add tensor trait for softmax()
